BIG changes, see desc
A song with more than one artist now displays all of them instead of just one.
A song with a known source will now display the source in the song title (needs testing to see if source is displayed correctly).

Fixed #9 by loading the music player when the user presses play instead of when the document has loaded (needs to be tested to see if the fix doesn't cause desync after a few minutes of playing). The fix was to set the audio src to empty at first and change it when the play button has been pressed.

Changed fonts from Lato to Open Sans.

Moved jQuery code to the JS file. 

Replaced in-line JS in the music player buttons with event listeners.